ECS-4 #comment Modifications suggested by Andreas and commited separately to Bitbucket. This integrates them in the version on the development branch. Original commit message:
The post processing is now taking place inside growcellsfromseeds3.cc which solves the problem of remaining unallocated pixels. The comment about the flag is no longer necessary.
I think the growing from seeds works pretty well now. We probably won't have to fix the bleeding into the background at this stage as we are not interested in the very edge of the cluster.